move test for whitespace to its own case
