+++ changedetectionio/tests/test_ignore_text.py
@@ -29,28 +29,6 @@ def test_strip_text_func():
     assert b"Some content" in stripped_content
 
 
-# Should be the same as set_original_ignore_response() but with a little more whitespacing
-def set_original_ignore_response_but_with_whitespace():
-    test_return_data = """<html>
-       <body>
-     Some initial text</br>
-     <p>
-     
-     
-     Which is across multiple lines</p>
-     <br>
-     </br>
-     
-         So let's see what happens.  </br>
-         
-         
-     </body>
-     </html>
-
-    """
-    with open("test-datastore/endpoint-content.txt", "w") as f:
-        f.write(test_return_data)
-
 def set_original_ignore_response():
     test_return_data = """<html>
        <body>
@@ -258,56 +236,3 @@ def test_check_global_ignore_text_functionality(client, live_server):
 
     res = client.get(url_for("api_delete", uuid="all"), follow_redirects=True)
     assert b'Deleted' in res.data
-
-# If there was only a change in the whitespacing, then we shouldnt have a change detected
-def test_check_ignore_whitespace(client, live_server):
-    sleep_time_for_fetch_thread = 3
-
-    # Give the endpoint time to spin up
-    time.sleep(1)
-
-    set_original_ignore_response()
-
-    # Goto the settings page, add our ignore text
-    res = client.post(
-        url_for("settings_page"),
-        data={
-            "minutes_between_check": 180,
-            "ignore_whitespace": "y",
-            'fetch_backend': "html_requests"
-        },
-        follow_redirects=True
-    )
-    assert b"Settings updated." in res.data
-
-    # Add our URL to the import page
-    test_url = url_for('test_endpoint', _external=True)
-    res = client.post(
-        url_for("import_page"),
-        data={"urls": test_url},
-        follow_redirects=True
-    )
-    assert b"1 Imported" in res.data
-
-    # Trigger a check
-    client.get(url_for("api_watch_checknow"), follow_redirects=True)
-
-    # Give the thread time to pick it up
-    time.sleep(sleep_time_for_fetch_thread)
-
-    # It should report nothing found (no new 'unviewed' class)
-    res = client.get(url_for("index"))
-    assert b'unviewed' not in res.data
-    assert b'/test-endpoint' in res.data
-
-    set_original_ignore_response_but_with_whitespace()
-
-    # Trigger a check
-    client.get(url_for("api_watch_checknow"), follow_redirects=True)
-
-    # Give the thread time to pick it up
-    time.sleep(sleep_time_for_fetch_thread)
-
-    # It should report nothing found (no new 'unviewed' class)
-    res = client.get(url_for("index"))
-    assert b'unviewed' not in res.data
